# brew.sh: reorder to speedup shellenv.