Reorder and restructure this file to make it as fast as possible for e.g. `brew --prefix`. Once we have that, also give `brew shellenv` short-circuit command execution to make it as fast as possible.

On my machine (with a warm cache) this gets:

```
brew shellenv  0.01s user 0.03s system 100% cpu 0.038 total
```

Compared to:

```
brew shellenv  0.09s user 0.19s system 100% cpu 0.277 total
```

before this commit.

Credit to @metacolin for giving me this idea in https://github.com/Homebrew/brew/pull/8846#discussion_r502982685

- [x] Have you followed the guidelines in our [Contributing](https://github.com/Homebrew/brew/blob/HEAD/CONTRIBUTING.md) document?
- [x] Have you checked to ensure there aren't other open [Pull Requests](https://github.com/Homebrew/brew/pulls) for the same change?
- [x] Have you added an explanation of what your changes do and why you'd like us to include them?
- [ ] Have you written new tests for your changes? [Here's an example](https://github.com/Homebrew/brew/blob/HEAD/Library/Homebrew/test/PATH_spec.rb).
- [x] Have you successfully run `brew style` with your changes locally?
- [x] Have you successfully run `brew tests` with your changes locally?
- [x] Have you successfully run `brew man` locally and committed any changes?

-----